+++ fractals/src/main/java/me/catmousedog/fractals/data/FractalValue.java
@@ -1,5 +1,15 @@
 package me.catmousedog.fractals.data;
 
+import me.catmousedog.fractals.paneloperators.fractals.Fractal;
+import me.catmousedog.fractals.paneloperators.functions.Function;
+
+/**
+ * A value returned by the {@link Fractal #get(double, double)} method. This
+ * <code>FractalValue</code> is then passed on to the current {@link Function}.
+ * <p>
+ * If the {@link FractalValue#i} equals the {@link FractalValue#I} the point
+ * associated with the <code>FractalValue</code> is convergent.
+ */
 public class FractalValue {
 
 	public double x, y;
@@ -19,6 +29,13 @@ public class FractalValue {
 		this.y = y;
 		this.i = i;
 		this.I = I;
+	}
+
+	/**
+	 * @return true if {@link FractalValue#i} equals {@link FractalValue#I}.
+	 */
+	public boolean isConvergent() {
+		return i == I;
 	}
 
 	/**