+++ fractals/src/main/java/me/catmousedog/fractals/paneloperators/fractals/BurningShip.java
@@ -41,7 +41,7 @@ public class BurningShip extends Fractal {
 			y = Math.abs(2 * tx * y) + cy;
 
 		}
-		return new FractalValue(0, 0, iterations, iterations);
+		return new FractalValue(x, y, iterations, iterations);
 	}
 
 	public @NotNull String informalName() {