+++ fractals/src/main/java/me/catmousedog/fractals/paneloperators/fractals/Multibrot.java
@@ -69,7 +69,7 @@ public class Multibrot extends Fractal {
 			x = C * Math.cos(t) + cx;
 			y = C * Math.sin(t) + cy;
 		}
-		return new FractalValue(0, 0, iterations, iterations);
+		return new FractalValue(x, y, iterations, iterations);
 	}
 
 	@Override